Added support for the offset request.
OffsetRequest seems to be basically working from the Consumer class, but it behaves somewhat erratically at the moment.  Correct number of bytes are being returned, but doesn't always contain the offset value expected(sometimes just comes back as zero and other times with what appears to be the offset)...bah.